requirements.txt
alabaster==0.7.9
aniso8601==1.2.0
argcomplete==1.9.2
asn1crypto==0.24.0
autodoc==0.3
Babel==2.3.4
base58==0.2.4
beautifulsoup4==4.5.3
boto3==1.7.10
botocore==1.10.10
cffi==1.9.1
cfn-flip==1.0.3
click==6.7
cryptography==2.1.4
decorator==4.0.11
docutils==0.13.1
durationpy==0.5
Flask==0.12
Flask-Ask==0.8.7
future==0.16.0
hjson==3.0.1
idna==2.2
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
MarkupSafe==0.23
nltk==3.2.5
Pillow==5.1.0
placebo==0.8.1
pyasn1==0.2.1
pycparser==2.17
Pygments==2.2.0
pyOpenSSL==16.2.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2016.10
PyYAML==3.12
requests==2.13.0
s3transfer==0.1.13
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.5.2
sphinx-autodoc-typehints==1.1.0
sphinx-rtd-theme==0.1.9
toml==0.9.4
tqdm==4.19.1
troposphere==2.2.1
Unidecode==1.0.22
waitress==1.0.2
WebOb==1.7.1
WebTest==2.0.25
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.45.1
